+++ Woof/Woof/Model/Rating/Stars.swift
@@ -1,0 +1,17 @@
+/// The units of measure in a 5-star rating.
+enum Stars: Int {
+    /// Value for one star.
+    case oneStar = 1
+
+    /// Value for two star.
+    case twoStars
+
+    /// Value for three star.
+    case threeStars
+
+    /// Value for four star.
+    case fourStars
+
+    /// Value for five star.
+    case fiveStars
+}